requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
crispy-bootstrap5==0.7
distlib==0.3.7
Django==4.2.3
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-filter==23.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
filelock==3.12.2
idna==3.4
Pillow==10.0.0
platformdirs==3.9.1
psycopg2==2.9.9
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.6
virtualenv==20.24.2
